+++ src/frontend/src/pages/HomePage/index.tsx
@@ -5,29 +5,21 @@ import { useRoom } from '@/hooks/queries/useRoom';
 import { useEffect, useState } from 'react';
 import { RoomDto } from '@/api/endpoints/room/room.interface';
 import { HomeSkeleton } from './HomeSkeleton';
+import { useDelayedLoading } from '@/hooks/utils/useDelayedLoading';
 
 export const HomePage = () => {
   const navigate = useNavigate();
+  const { getRooms } = useRoom();
+  const showSkeleton = useDelayedLoading(getRooms.data);
   const [videos, setVideos] = useState<RoomDto[]>([]);
-  const [showSkeleton, setShowSkeleton] = useState(true);
 
   const handleVideoCardClick = (code: string) => {
     navigate(`/room?code=${code}`);
   };
 
-  const { getRooms } = useRoom();
-
   useEffect(() => {
-    const minLoadingTime = 300;
-    const startTime = Date.now();
-
     if (getRooms.data) {
       setVideos(getRooms.data.pages.flatMap(page => page));
-
-      const elapsedTime = Date.now() - startTime;
-      const delay = Math.max(0, minLoadingTime - elapsedTime);
-
-      setTimeout(() => setShowSkeleton(false), delay);
     }
   }, [getRooms.data]);
